requirements.txt
pandas>=0.19.1
Biopython>=1.6
